requirements.txt
aiofiles==0.6.0
aiomysql==0.0.21
alembic==1.5.5
argon2-cffi==20.1.0
beautifulsoup4==4.9.3
databases==0.4.1
email-validator==1.1.2
fastapi==0.63.0
fastapi-utils==0.2.1
httpx==0.16.1
Jinja2==2.11.3
PyMySQL==0.9.3
mysqlclient==2.0.3
python-jose==3.2.0
python-multipart==0.0.5
passlib==1.7.4
pydantic==1.8
python-dotenv==0.15.0
superslug==1.0.1
typer==0.3.2
uvicorn==0.13.4
